Rfcomm channel has uint8_t type (Marcel Holtmann)

+++ include/devices/unixbluetooth.h
@@ -44,7 +44,7 @@
 
 #include "misc.h"
 
-int bluetooth_open(bdaddr_t *bdaddr, int channel, struct gn_statemachine *state);
+int bluetooth_open(bdaddr_t *bdaddr, uint8_t channel, struct gn_statemachine *state);
 int bluetooth_close(int fd, struct gn_statemachine *state);
 int bluetooth_write(int fd, const __ptr_t bytes, int size, struct gn_statemachine *state);
 int bluetooth_read(int fd, __ptr_t bytes, int size, struct gn_statemachine *state);
